cs4265-social-media-trends/requirements.txt
pandas
pyyaml
# Added for future Spark/HDFS work
pyspark
python-dotenv